+++ packages/plugin-dev-changesets/src/commands/changesets/create.ts
@@ -1,7 +1,7 @@
 import {
   BaseCommand,
   ContextCommand,
-  Flags,
+  Oclif,
   Interfaces as _Interfaces,
 } from "@business-as-code/core";
 
@@ -16,28 +16,28 @@ hello friend from oclif! (./src/commands/hello/index.ts)
   ];
 
   static override flags = {
-    // message: Flags.string({
+    // message: Oclif.Flags.string({
     //   description: "Workspace name",
     //   required: true,
     // }),
     /** Moon scopes - https://tinyurl.com/2ek7rph4 . @todo - validate this better here */
-    query: Flags.string({
+    query: Oclif.Flags.string({
       description: "Query to select snapshotted projects",
       required: false,
       default: "projectType=library || projectType=application",
     }),
-    message: Flags.string({
+    message: Oclif.Flags.string({
       description: "Message for the changeset",
       required: true,
     }),
-    bump: Flags.custom<"major" | "minor" | "patch">({
+    bump: Oclif.Flags.custom<"major" | "minor" | "patch">({
       summary: "Specify level for bumping",
       options: ["major", "minor", "patch"],
       helpGroup: "GLOBAL",
       default: "patch",
       required: true,
     })(),
-    // bump: Flags.custom({
+    // bump: Oclif.Flags.custom({
     //   parse: async (input: string, opts) => {
     //     if (["major", "minor", "patch"].includes(input)) {
     //       return input;
@@ -45,21 +45,21 @@ hello friend from oclif! (./src/commands/hello/index.ts)
     //   },
     // }),
 
-    workspacePath: Flags.string({
+    workspacePath: Oclif.Flags.string({
       description: "Workspace name",
       required: false,
     }),
-    // configPath: Flags.string({
+    // configPath: Oclif.Flags.string({
     //   description: "Relative or absolute path to a workspace configuration",
     //   required: false,
     // }),
-    // cliVersion: Flags.string({
+    // cliVersion: Oclif.Flags.string({
     //   description: "Specify a Bac cli version",
     //   required: false,
     //   default: "*",
     // }),
 
-    // registry: Flags.string({
+    // registry: Oclif.Flags.string({
     //   description: "Specify a package manager registry to load the Bac cli",
     //   required: false,
     //   default: "http://localhost:4873",
@@ -67,7 +67,7 @@ hello friend from oclif! (./src/commands/hello/index.ts)
   };
 
   static override args = {
-    // path: Args.string({
+    // path: Oclif.Args.string({
     //   description: "Absolute/Relative path",
     //   required: false,
     // }),