+++ packages/plugin-core-essentials/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @business-as-code/plugin-core-essentials
 
+## 0.0.0-latest-20230811110927
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230811110927
+  - @business-as-code/core@0.0.0-latest-20230811110927
+  - @business-as-code/error@0.0.0-latest-20230811110927
+  - @business-as-code/fslib@0.0.0-latest-20230811110927
+
+## 0.0.0-latest-20230811110816
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230811110816
+  - @business-as-code/core@0.0.0-latest-20230811110816
+  - @business-as-code/error@0.0.0-latest-20230811110816
+  - @business-as-code/fslib@0.0.0-latest-20230811110816
+
+## 0.0.0-latest-20230811110751
+
+### Patch Changes
+
+- this is a snapshot release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20230811110751
+  - @business-as-code/core@0.0.0-latest-20230811110751
+  - @business-as-code/error@0.0.0-latest-20230811110751
+  - @business-as-code/fslib@0.0.0-latest-20230811110751
+
 ## 0.0.0-latest-20230811071401
 
 ### Patch Changes